requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
redis==5.0.8
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
